+++ packages/ink_widget/CHANGELOG.md
@@ -1,3 +1,3 @@
 ## 1.0.0
 ### Added
-* Added `InkWellWidget` wrapper widget over InkWell
+* Added `InkWidget` wrapper widget over InkWell